Path of file to be read/written to/from needs to be changed
Right now the path is set to a file named lib in my documents, something
like C:\Users\Alexr0\Documents\lib. You're gonna have to change it if
you want it to work on your computer.
I put pretty long comments that look like this:
//**********************NEED TO CHANGE THIS

directly above the places that need the change