Fixed the "save" issues!
No longer deletes the old song and replaces with a new song, now it just
edits the text. The remove/add way would be bad for sorting I realized.
Also, I fixed the duplicate bug. Now when you edit, you can't have 2
songs with the same artist+title but different info. Additionally, you
can now edit a song's year + album. Before it was seeing that a song
with the same title/artist existed and it wouldn't allow it.

All in all, that's all sorted out. Now we just need to get to sorting :P